refactor: moved UserResponse class
