+++ app-bpm/app/src/main/java/com/contextaware/app_bpm/data/network/AuthApiService.kt
@@ -1,6 +1,6 @@
 package com.contextaware.app_bpm.data.network
 
-import com.google.gson.annotations.SerializedName
+import com.contextaware.app_bpm.data.model.UserResponse
 import retrofit2.Response
 import retrofit2.http.Header
 import retrofit2.http.POST
@@ -12,11 +12,3 @@ interface AuthApiService {
     @POST("users/signup")
     suspend fun signupUser(@Header("Authorization") authHeader: String): Response<UserResponse>
 }
-
-data class UserResponse(
-    val id: Int,
-    val username: String?,
-    val email: String,
-    @SerializedName("created_at")
-    val createdAt: String?,
-)
